Moved version check out of `upToDateWhen` block. Added  annotationsFile into task input to include it into up-to-date check. Added tests for multiple Gradle versions.
