+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/kotlin/org/jetbrains/kotlin/gradle/KotlinGradlePluginIT.kt
@@ -428,37 +428,4 @@ class KotlinGradleIT: BaseGradleIT() {
             assertSuccessful()
         }
     }
-
-    @Test
-    fun testGradleJavaIcWorking() {
-        val project = Project("kotlinJavaProject", "2.14.1")
-        project.setupWorkingDir()
-
-        val buildScript = File(project.projectDir, "build.gradle")
-
-        buildScript.modify { "$it\n" + "compileJava.options.incremental = true" }
-        project.build("build") {
-            assertSuccessful()
-        }
-
-        // Then modify a Java source and check that compileJava is incremental:
-        File(project.projectDir, "src/main/java/demo/HelloWorld.java").modify { "$it\n" + "class NewClass { }" }
-        project.build("build") {
-            assertSuccessful()
-            assertNotContains("not incremental")
-            assertContains("Incremental compilation")
-        }
-
-        // Then modify a Kotlin source and check that Gradle sees that Java is not up-to-date:
-        File(project.projectDir, "src/main/kotlin/helloWorld.kt").modify {
-            it.trim('\r', '\n').trimEnd('}') + "\nval z: Int = 0 }"
-        }
-        project.build("build") {
-            assertSuccessful()
-            assertContains(":compileKotlin")
-            assertNotContains(":compileJava UP-TO-DATE")
-            assertNotContains("not incremental")
-            assertNotContains("None of the classes needs to be compiled!")
-        }
-    }
 }
