+++ libraries/tools/kotlin-gradle-plugin/src/main/kotlin/org/jetbrains/kotlin/gradle/plugin/KotlinPlugin.kt
@@ -555,19 +555,21 @@ private fun configureJavaTask(kotlinTask: KotlinCompile, javaTask: AbstractCompi
     // Gradle Java IC in older Gradle versions (before 2.14) cannot check .class directories updates.
     // To make it work, reset the up-to-date status of compileJava with this flag.
     kotlinTask.anyClassesCompiled = false
-    javaTask.outputs.upToDateWhen { task ->
-        val gradleSupportsJavaIcWithClassesDirs = ParsedGradleVersion.parse(javaTask.project.gradle.gradleVersion)
-                                                          ?.let { it >= ParsedGradleVersion(2, 14) } ?: false
-        val kotlinClassesCompiled = kotlinTask.anyClassesCompiled
-        if (!gradleSupportsJavaIcWithClassesDirs && kotlinClassesCompiled) {
-            logger.info("Marking $task out of date, because kotlin classes are changed")
-            return@upToDateWhen false
+    val gradleSupportsJavaIcWithClassesDirs = ParsedGradleVersion.parse(javaTask.project.gradle.gradleVersion)
+                                                      ?.let { it >= ParsedGradleVersion(2, 14) } ?: false
+    if (!gradleSupportsJavaIcWithClassesDirs) {
+        javaTask.outputs.upToDateWhen { task ->
+            if (kotlinTask.anyClassesCompiled) {
+                logger.info("Marking $task out of date, because kotlin classes are changed")
+                false
+            } else true
         }
-        true
     }
 
-    // Make Gradle check if the javaTask up-to-date based on the Kotlin classes
+    // Make Gradle check if the javaTask is up-to-date based on the Kotlin classes
     javaTask.inputs.dir(kotlinTask.destinationDir)
+    // Also, use kapt1 annotations file for up-to-date check since annotation processing is done with javac
+    kotlinTask.kaptOptions.annotationsFile?.let { javaTask.inputs.file(it) }
 
     javaTask.dependsOn(kotlinTask.name)
     /*
